syncbase: Fix memory leak in runRemoteOp.
The runRemoteOp function takes a context and creates a child context.
If the rpc operation is successful, the child context is not cancelled,
and the caller is implicity required to cancel the parent.

In the clock code path, the context passed to runRemoteOp is never
cancelled, and so the child contexts accumulate.

This CL makes the clock create a new context with cancel, and cancels it
in a defer.

Change-Id: I853cb4eb4bd9de35e7be6f15c75cd0940fbfc24c