syncbase/bridge/cgo: Two changes to the C API for handling callbacks
The first change is to switch the handle from 'int' to 'void*'. This
will help interfacing with language where is easier to pass a pointer
to a struct instead of explicitly managing the mapping to it. Both
Java and Swift manage memory in a way that makes is hard/impossible
to use pointers so the 'void*' will be used as a larger 'int'.

The second change is to use a single handle instead of two for both
watching and scanning. This makes the C API simpler and is more inline
with what the Java/Swift need to do (namely, keeping a struct that
describes both callbacks).

Note: the author of the Swift code is Aaron Zinman <aaron@azinman.com>.

MultiPart: 1/2
Change-Id: I6aa1f300a7d97aa1c9ef0f20b08cde7ac29cd5fa