java: Implement the Database.WatchPatterns and Collection.Scan
I wasn't able to see that the onChange is called (which matches what
zinman@ also observed).

This change also contains two fixes to the ScanCallbacks interface:

 - the onValue method is renamed to onKeyValue
 - the onDone is updated to take a VError argument.

Due to some new "inconsistent definitions" I also had to move the
jFindClass from jni_util to jni.go. The bottom of the jni.go and the
jni_lib.go are also sorted.

MultiPart: 1/2
Change-Id: Ie3a07a22cadb88fe3c888da5453fbc3ff0e293f1