services/mounttable/btmtd: Keep track of nodes per user
Keep track of who created each node and how many they've created. This
is only accounting. There is no quota yet.

We use the same logic as mounttabled to identify the creator of a node.

Each node now contains the name of its creator, and we store one counter
per creator in a separate table. The counter is incremented when a node
is created, and decremented when a node is deleted.

Change-Id: I8a0045c393b00cf48aaaec76dbf08ff2d9cd2ccd