services/mounttable/btmtd: Keep track of mounted servers
Add a counter for the number of mounted servers, per user. Each server
is charged to the creator of the node where it is mounted.

Since we're keeping track of the mounted servers, we can't use
bigtable's garbage collection to delete expired servers. Instead, server
expiration is now done in gc().

Change-Id: Ia1b1b59ff68d1bf933828c975bc6ca7053c92bea