x/ref: ListenForAppPeers and AdvertiseApp
Add functions to syncbase's discovery library that allow one to
advertise apps and listen for peers of that app too.

The flexibility is heavily locked down. This current cut only
allows advertising and scanning on the app that your default
blessing string has (according to our blessings convention).
e.g., dev.v.io:o:<app>:<user> is only allowed to advertise
their presence in <app> and scan for other users with <app>.

Note: This is only an API limitation, not a security restriction.

Change-Id: I68c3f9a369e7cc0e67dbc972e25a6d01f1beb9cb