java: Get the DatabaseOptions.rootDir to work
Note that the rootDir is not exposed in the io.v.syncbase.core. The
purpose of this CL is to get all the high-level API tests from
SyncbaseTest to work. This change also adds the ability to shutdown
syncbase. The high-level API tests are reconfigured to use and empty
list of mounts points speed up the shutdown.

MultiPart: 1/3
Change-Id: Iad82088f94a5dc9c0304155648d87df15b88bc55