syncbase cgo: Fix the compilation warnings on Android
The jni.h is slightly different on Android than a local machine. More
specific, in Android the AttachCurrentThread and AttachCurrentThreadAsDaemon
take as a second argument a JNIEnv** while on the local machine is
void**. This change places those two functions in .c files that use Go
build tags to restrict the compilation to the appropriate
environments.

Change-Id: I6879444837c1e9582923859cdb1d4dc403c16de4