x/ref: syncbase cgo bridge ListenForAppPeers and ListenForInvites
We use the syncbase/discovery implementations of ListenForAppPeers
and ListenForInvites to surface them in the cgo bridge.

The next steps are to plumb this through in jni (and Swift).

Note: Untested. All I can tell is that it compiles. Other than doing
the JNI + high-level implementation, how can I check?

Note2: I've been hooking this up through JNI and high-level Java.
It compiles there and has approximately the correct format.

Change-Id: I00c6ff045073cc7c3d9b32769cf3f2858b7bd0f3