services/mounttable/btmtd: Add row id
Before this change, we implicitly relied on createChild() never being
called concurrently within the same millisecond. In reality, this can
happen and does happen once in a while in the tests.

With this change, each row has its own 4-byte ID and the parent uses
both the child's ID and its name to reference it.

(This is not backward compatible)

Change-Id: Ib0334a25caa98307cf1eeed4932454c5721310cc