x/ref: Avoid JNI DETECTED ERROR IN APPLICATION error
I noticed that if I try to return a non-primitive object during an
exception (such as during SyncgroupJoin) that the jni would crash.

This error seems to imply that we have to wait for Java to handle
the exception OR clear the exception OR just return early (with null).

I opted for returning early. The JNI functions that return errors and
primitives remain as-is since the latter can't be nil'd.

Change-Id: If7d88ee72cead5a6f0c0d38061a2f309570089d3