syncbase: Allow Syncbase root dir to be moved
When Syncbase creates a new service instance, it stores the location
of the new LevelDB as an absolute path (DbInfo.RootDir). If Syncbase's
root directory is ever moved, it won't be able to boot regardless of
the updated rootDir parameter passed at init (which was ignored). This
CL changes this behavior to store relative paths in DbInfo, while
allowing existing absolute paths to continue to work. This CL does not
correct the older values, which means this CL only fixes the behavior
for new Syncbase instances.

MultiPart: 2/2
Change-Id: I42721ce61b8186cb95849e2d63752086aca50271