v.io/x/ref/services/syncbase: start ServerBlobFetcher() in server,NewService()
This change moves the starting of the vsync.ServerBlobFetcher()
back into server,NewService(), and wait for it to exit on service.Close().
I've checked (with printfs, now removed) that this is indeed being invoked by some
existing tests.

Some of the prerequisites for the blob fetcher are not set up in tests of
vsync, which was causing some error messages to be generated.

Change-Id: I18e969924bf3a085af602fc8bf71603ec6b22b1c