x/ref: Separate impl.go and jni.go ptr maps
The reason they need to be separate is that for inviteScan and
peerScan, we are using the same handle value.

ptr maps do not accept duplicate keys, and they no longer generate
unique keys for us. Thus, it is better if we use separate maps in
the jni-side and impl-side.

Change-Id: I4d892471583d5be90637c330d619fed1bb0ba59f